+++ examples/classic-es5/invalid/optional.server.js
@@ -1,0 +1,2 @@
+var assignmentGroup = "97c04b3b1b12100043ab85e5bd0713e2";
+current.assignment_group = assignmentGroup;